Address some of the review comments
