+++ certbot/interfaces.py
@@ -604,204 +604,3 @@ class RenewDeployer(object):
         :type lineage: storage.RenewableCert
 
         """
-
-
-@six.add_metaclass(abc.ABCMeta)
-class ParserNode(object):
-    """
-    ParserNode is the basic building block of the tree of such nodes,
-    representing the structure of the configuration. It is largely meant to keep
-    the structure information intact and idiomatically accessible.
-
-    The root node as well as the child nodes of it should be instances of ParserNode.
-    Nodes keep track of their differences to on-disk representation of configuration
-    by marking modified ParserNodes as dirty to enable partial write-to-disk for
-    different files in the configuration structure.
-
-    While for the most parts the usage and the child types are obvious, "include"-
-    and similar directives are an exception to this rule. This is because of the
-    nature of include directives - which unroll the contents of another file or
-    configuration block to their place. While we could unroll the included nodes
-    to the parent tree, it remains important to keep the context of include nodes
-    separate in order to write back the original configuration as it was.
-    """
-
-    @abc.abstractmethod
-    def arguments(self):
-        """
-        Returns a list of arguments for this ParserNode object
-
-        :returns: list of arguments
-        """
-
-    @abc.abstractmethod
-    def dirty(self):
-        """
-        Returns a boolean telling if the state of the node was changed after the
-        initial read or the last save operation. The main usage for this method is
-        to discover the nodes in the tree that need to be written to disk upon
-        save.
-
-        :param str name: The name of the child node to add
-
-        :returns: True if the node needs to be written to disk
-        """
-
-    @abc.abstractmethod
-    def save(self, msg):
-        """
-        Save traverses the children, and attempts to write the AST to disk for
-        all the objects that are marked dirty. The actual operation of course
-        depends on the underlying implementation. save() shouldn't be called
-        from the Configurator outside of it's designated save() method in order
-        to ensure that the Reverter checkpoints are created properly.
-
-        Note: this approach of keeping internal structure of the configuration
-        within the ParserNode tree does not represent the file inclusion structure
-        of actual configuration files that reside in the filesystem. To handle
-        file writes properly, the file specific temporary trees should be extracted
-        from the full ParserNode tree where necessary when writing to disk.
-
-        """
-
-    @abc.abstractmethod
-    def set_arguments(self, arguments):
-        """
-        Sets argument list of the ParserNode object and marks the node dirty.
-        """
-
-
-@six.add_metaclass(abc.ABCMeta)
-class CommentNode(ParserNode):
-    """
-    CommentNode class is used for representation of comments within the parsed
-    configuration structure. Because of the nature of comments, it is not able
-    to have child nodes and hance it is always treated as a leaf node.
-    """
-
-
-@six.add_metaclass(abc.ABCMeta)
-class DirectiveNode(ParserNode):
-    """
-    DirectiveNode class represents a configuration directive within the configuration.
-    It can have zero or more arguments attached to it. Because of the nature of
-    single directives, it is not able to have child nodes and hence it is always
-    treated as a leaf node.
-    """
-
-
-@six.add_metaclass(abc.ABCMeta)
-class BlockNode(ParserNode):
-    """
-    BlockNode class represents a block of nested configuration directives, comments
-    and other blocks as its children. A BlockNode can have zero or more arguments
-    attached to it.
-
-    Configuration blocks typically consist of one or more child nodes of all possible
-    types. Because of this, the BlockNode class has various discovery and structure
-    management methods.
-    """
-
-    @abc.abstractmethod
-    def add_child_block(self, name, arguments=[], position=-1):
-        """
-        Adds a new BlockNode child node with provided values and marks the callee
-        BlockNode dirty. This is used to add new children to the AST.
-
-        :param str name: The name of the child node to add
-        :param list arguments: list of arguments for the node
-        :param int position: Position in the list of children to add the new child
-            node to. Defaults to -1, which appends the newly created node to the list.
-
-        :returns: BlockNode instance of the created child block
-
-        """
-
-    @abc.abstractmethod
-    def add_child_directive(self, name, arguments=[], position=-1):
-        """
-        Adds a new DirectiveNode child node with provided values and marks the
-        callee BlockNode dirty. This is used to add new children to the AST.
-
-        :param str name: The name of the child node to add
-        :param list arguments: list of arguments for the node
-        :param int position: Position in the list of children to add the new child
-            node to. Defaults to -1, which appends the newly created node to the list.
-
-        :returns: DirectiveNode instance of the created child directive
-
-        """
-
-    @abc.abstractmethod
-    def add_child_comment(self, name, arguments=[], position=-1):
-        """
-        Adds a new CommentNode child node with provided values and marks the
-        callee BlockNode dirty. This is used to add new children to the AST.
-
-        :param str name: The name of the child node to add
-        :param list arguments: list of arguments for the node
-        :param int position: Position in the list of children to add the new child
-            node to. Defaults to -1, which appends the newly created node to the list.
-
-        :returns: CommentNode instance of the created child comment
-
-        """
-
-    @abc.abstractmethod
-    def find_blocks(self, name, arg=None):
-        """
-        Find a configuration block with name and (optionally) an argument.
-        This method walks the child tree of ParserNodes under the instance it was
-        called from. This way it is possible to search for the whole configuration
-        tree, when starting from root node or to do a partial search when starting
-        from a specified branch.
-
-        :param str name: The name of the directive to search for
-        :param str arg: Return only directives that have argument with provided value
-
-        :returns: A list of found BlockNode objects.
-        """
-
-
-    @abc.abstractmethod
-    def find_directives(self, name, arg=None):
-        """
-        Find a directive with name and (optionally) an argument. This method walks the
-        child tree of ParserNodes under the instance it was called from. This way
-        it is possible to search for the whole configuration tree, when starting
-        from root node, or to do a partial search when starting from a specified
-        branch.
-
-        :param str name: The name of the directive to search for
-        :param str arg: Return only directives that have argument with provided value
-
-        :returns: A list of found DirectiveNode objects.
-
-        """
-
-    @abc.abstractmethod
-    def find_comments(self, comment, exact=False):
-        """
-        Find comments with value containing or being exactly the same as search term.
-
-        This method walks the child tree of ParserNodes under the instance it was
-        called from. This way it is possible to search for the whole configuration
-        tree, when starting from root node, or to do a partial search when starting
-        from a specified branch.
-
-        :param str comment: The content of comment to search for
-        :param bool exact: If the comment needs to exactly match the search term
-
-        :returns: A list of found CommentNode objects.
-
-        """
-
-    @abc.abstractmethod
-    def delete_child(self, child):
-        """
-        Remove a specified child node from the list of children of the called
-        BlockNode object.
-
-        :param ParserNode child: Child ParserNode object to remove from the list
-            of children of the callee.
-        """
